HotFix/Off-ramp fixed

+++ backend/cb_flow_connect/src/main/java/com/cb_labs/cb_flow_connect/web/dto/models/capa/response/CapaQuotingResponse.java
@@ -12,7 +12,7 @@ public record CapaQuotingResponse(
         String fiatCurrency,
         String blockchainSymbol,
         String tokenSymbol,
-        Long cryptoAmount,
+        Double cryptoAmount,
         Double rate,
         String flow
     ) {}
