+++ backend/cb_flow_connect/src/main/java/com/cb_labs/cb_flow_connect/web/dto/response/QuotingResponse.java
@@ -2,5 +2,5 @@ package com.cb_labs.cb_flow_connect.web.dto.response;
 
 public record QuotingResponse(
     Double fiatAmount,
-    Long cryptoAmount
+    Double cryptoAmount
 ) { }
